Try to add middleware
