+++ backend/src/entities/preference.entity.ts
