+++ backend/src/login/dtos/login-form-dto.ts
@@ -1,4 +1,5 @@
 export class LoginFormDto {
   email: string;
   password: string;
+  type: string = "";
 }
