+++ backend/src/middlewares/auth.middleware.ts
@@ -1,0 +1,15 @@
+import { Injectable, NestMiddleware } from '@nestjs/common';
+import { Request, Response, NextFunction } from 'express';
+
+export class AuthUserGetter {
+  
+}
+
+@Injectable()
+export class AuthMiddleware implements NestMiddleware {
+  use(req: Request, res: Response, next: NextFunction) {
+    
+    console.log('Request...');
+    next();
+  }
+}
