+++ backend/src/register/register.module.ts
@@ -5,7 +5,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserAuthRepository } from './repositories/user-auth.repository';
 import { TrainerProfileRepository } from './repositories/trainer-profile.repository';
 import { TraineeProfileRepository } from './repositories/trainee-profile.repository';
-import { Preference } from './entities/preference.entity';
+import { Preference } from '../entities/preference.entity';
 @Module({
   imports: [
     TypeOrmModule.forFeature([
